requirements.txt
# Core Python
numpy>=1.23
scipy>=1.10

# PDF processing
pdfplumber>=0.10.2
pypdf>=4.0.0

# NLP & text processing
nltk>=3.8.1
spacy>=3.7.2

# Sentence embeddings
sentence-transformers>=2.6.1
scikit-learn>=1.3.2

# Knowledge Graph
networkx>=3.2.1
python-louvain>=0.16

# Vector storage (local)
faiss-cpu>=1.7.4
chromadb>=0.4.24

# LLM integration (local)
ollama>=0.1.7
langchain>=0.1.16
langchain-community>=0.0.34

# Config & utilities
pyyaml>=6.0.1
tqdm>=4.66.2
loguru>=0.7.2
